Complete very basic proof-of-concept
Currently, the extension runs from a 'test.html' popup instead of the actual GUI, as the interface has not been updated to meet the new functionality as of yet. Furthermore, scripts do not yet run automatically, but instead only when the 'enable script' button in the popup is clicked. Storage has, however, been implemented.